ref(eos): <- rename incremerkle mod for clarity
