+++ common/eos/src/core_initialization/eos_init_utils.rs
@@ -14,7 +14,7 @@ use crate::{
     eos_crypto::eos_private_key::EosPrivateKey,
     eos_database_utils::EosDbUtils,
     eos_global_sequences::ProcessedGlobalSequences,
-    eos_merkle_utils::Incremerkle,
+    eos_incremerkle::Incremerkle,
     eos_producer_schedule::EosProducerScheduleV2,
     eos_submission_material::EosSubmissionMaterial,
     eos_types::{Checksum256s, EosBlockHeaderJson, EosKnownSchedules},
